pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wmark"
version = "0.0.1"
description = "Multi-bit LLM watermarking under model compression"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "transformers>=4.40",
    "accelerate>=0.30",
    "bitsandbytes>=0.43",
    "auto-gptq>=0.7",
    "autoawq>=0.2.5",
    "datasets>=2.18",
    "numpy>=1.26",
    "scipy>=1.12",
    "scikit-learn>=1.4",
    "pyldpc>=0.7.9",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "pandas>=2.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4"]
